celebamask_baseline/solution/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy
pillow
tqdm
albumentations>=1.4.0